test/requirements.txt
requests == 2.32.3
uuid == 1.30
Pillow == 10.4.0
